% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Groningen100}
\alias{Groningen100}
\title{Geographic data}
\format{
A data frame with five variables: \code{lon}, \code{lat},
  \code{som}, \code{naam} and \code{categorie}.
}
\usage{
Groningen100
}
\description{
Geographic data
}
\keyword{datasets}
